Remove misleading "bare" aliases
These make it seem like `openjdk:26` is GA, and we don't want that (in fact, we aggressively don't want that).

This also removes a lot of old cruft we haven't supported/used in ~3 years, such as Alpine support (last supported on jdk.java.net in 2022) and JRE support (last supported here in 2022 as well, with prior builds).